api-gateway/requirements.txt
django
djangorestframework
drf-yasg
spyne
requests